refactor(src): split metrics collection into modular units
- Move CPU, disk, memory, and network gathering into dedicated submodules
- Delete the monolithic metrics.rs implementation and keep payload assembly in src/metrics/mod.rs
- Add memory parsing tests and tighten MemAvailable fallback behavior
- Preserve non-Linux network handling while isolating the warning path
